+++ app.go
@@ -13,16 +13,16 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/zasper-io/zasper/analytics"
-	"github.com/zasper-io/zasper/content"
-	"github.com/zasper-io/zasper/core"
-	"github.com/zasper-io/zasper/gitclient"
-	"github.com/zasper-io/zasper/health"
-	"github.com/zasper-io/zasper/kernel"
-	"github.com/zasper-io/zasper/kernelspec"
-	"github.com/zasper-io/zasper/search"
-	"github.com/zasper-io/zasper/session"
-	"github.com/zasper-io/zasper/websocket"
+	"github.com/zasper-io/zasper/internal/analytics"
+	"github.com/zasper-io/zasper/internal/content"
+	"github.com/zasper-io/zasper/internal/core"
+	"github.com/zasper-io/zasper/internal/gitclient"
+	"github.com/zasper-io/zasper/internal/health"
+	"github.com/zasper-io/zasper/internal/kernel"
+	"github.com/zasper-io/zasper/internal/kernelspec"
+	"github.com/zasper-io/zasper/internal/search"
+	"github.com/zasper-io/zasper/internal/session"
+	"github.com/zasper-io/zasper/internal/websocket"
 
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
