+++ internal/content/content_manager.go
@@ -9,8 +9,8 @@ import (
 	"sort"
 	"time"
 
-	"github.com/zasper-io/zasper/core"
-	"github.com/zasper-io/zasper/models"
+	"github.com/zasper-io/zasper/internal/core"
+	"github.com/zasper-io/zasper/internal/models"
 
 	"github.com/rs/zerolog/log"
 )
